split project/samples_test
